+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,7 @@
+{
+  "ExpandedNodes": [
+    ""
+  ],
+  "SelectedNode": "\\PPAI_DSI_Grupo5.sln",
+  "PreviewInSolutionExplorer": false
+}
