+++ PPAI_DSI_Grupo5/Entidades/Negocio/AsignacionCientificoCI.cs
