+++ PPAI_DSI_Grupo5/Entidades/Negocio/AsignacionDirectorCI.cs
