+++ PPAI_DSI_Grupo5/Entidades/Negocio/Caracteristica.cs
