+++ PPAI_DSI_Grupo5/Entidades/Negocio/CaracteristicaRecurso.cs
