+++ PPAI_DSI_Grupo5/Presentacion/ABM Turno/AltaTurno.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace PPAI_DSI_Grupo5.Presentacion.ABM_Turno
+{
+    public partial class AltaTurno : Form
+    {
+        public AltaTurno()
+        {
+            InitializeComponent();
+        }
+
+        private void panel1_Paint(object sender, PaintEventArgs e)
+        {
+
+        }
+    }
+}
