+++ PPAI_DSI_Grupo5/Presentacion/Transacciones/RegistrarTurno.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace PPAI_DSI_Grupo5.Presentacion.Transacciones
+{
+    public partial class RegistrarTurno : Form
+    {
+        public RegistrarTurno()
+        {
+            InitializeComponent();
+        }
+    }
+}
